[refactor] fix builtin function libraries for SPIR

+++ include/opencl_math.cl
@@ -2,16 +2,16 @@
 /**
  * work-item related functions
  */
-ulong amp_get_global_id(ulong n) {
-  return (long)get_global_id((uint)n);
+ulong amp_get_global_id(uint n) {
+  return (long)get_global_id(n);
 }
 
-ulong amp_get_local_id(ulong n) {
-  return (long)get_local_id((uint)n);
+ulong amp_get_local_id(uint n) {
+  return (long)get_local_id(n);
 }
 
-ulong amp_get_group_id(ulong n) {
-  return (long)get_group_id((uint)n);
+ulong amp_get_group_id(uint n) {
+  return (long)get_group_id(n);
 }
 
 void amp_barrier(uint n) {
